@RealFYang 2023-11-14
/approval request Applies clean, low-risk, risc-v only change which improves MD5 intrinsic performance on linux-riscv platform a bit more and changes usage of scratch registers t0 and t1 so that will be safer. Correctness and performance tested on sifive unmatched.